requirements.txt
# Core
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
joblib>=1.2.0
scipy>=1.7.0

# Experimentation
Jinja2>=3.0.0

# MLflow (optional)
mlflow>=2.0.0

# API
flask>=2.0.0
Waitress>=2.1.0

# Dashboard
streamlit>=1.20.0

# Data storage (parquet - optional; falls back to CSV)
# pyarrow>=10.0.0

# Kafka (optional)
kafka-python>=2.0.2
jsonschema>=4.0.0

# Orchestration
apache-airflow>=2.5.0

# Dev & testing
pytest>=7.0.0
matplotlib>=3.5.0
